Auto-fixed SVG
Windows path were auto-fixed to Linux path by the renderfarm (no problem if you use Windows). This process is automatic in the renderfarm of Pepper&Carrot and is here to workaround an Inkscape bug. (the Linux version can't read the path made by Windows for reasons, but the way around works fine for reasons too.)